third_party/blink/web_tests/platform/mac/virtual/scalefactor200withoutzoom/external/wpt/largest-contentful-paint/observe-after-untrusted-scroll-expected.txt

This is a testharness.js-based test.
FAIL Same-origin image after a JS initiated scroll event is observable. assert_equals: expected 0 but got 21.780000000035216
Harness: the test ran to completion.
